Remove buefy.
Replace use of buefy, which isn't maintained any longer, with plain
bulma.

I'm sticking with bulma 0.9.4 since there doesn't seem to be much point
in trying to migrate to a new version; we gain very little in terms of
reduced complexity or maintainability but still have to invest a lot in
reaching parity.

Also adding a bunch of new tests here to cover all the components.